requirements.txt
openai
streamlit
fpdf
python-dotenv
python-docx